# internal/docs/swagger.yaml
definitions:
...
  api.CreateVariantRequest:
    properties:
      params:
        type: object
      type:
        type: string
    type: object
  api.CreateVariantResponse:
    properties:
      job_id:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.ListVariantsResponse:
    properties:
      rebuilding:
        type: boolean
      variants:
        items:
          $ref: '#/definitions/api.VariantResponse'
        type: array
    type: object
...
  api.VariantResponse:
    properties:
      asset_version_id:
        type: string
      created_at:
        type: string
      download_url:
        type: string
      id:
        type: string
      size:
        type: integer
      storage_key:
        type: string
      transform_params:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/assets/{id}/preview:
    get:
      description: 'Applies a resize/format transform to the asset''s current version
        in memory and returns the result directly (never stored). Responses are cached
        in an LRU cache (100 entries, 5-minute TTL) so repeated identical calls are
        cheap. <br>Query parameters: <ul> <li><strong>w</strong> — Target width in
        pixels</li> <li><strong>h</strong> — Target height in pixels</li> <li><strong>fit</strong>
        — Fit mode: <code>contain</code> (default), <code>cover</code>, <code>fill</code></li>
        <li><strong>format</strong> — Output format: <code>jpeg</code> (default),
        <code>png</code>, <code>webp</code></li> <li><strong>q</strong> — JPEG quality
        1–100 (default: encoder default)</li> </ul> Only supported for image assets.
        Returns 400 for video, audio, or PDF assets.'
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Target width
        in: query
        name: w
        type: integer
      - description: Target height
        in: query
        name: h
        type: integer
      - description: Fit mode (contain|cover|fill)
        in: query
        name: fit
        type: string
      - description: Output format (jpeg|png|webp)
        in: query
        name: format
        type: string
      - description: JPEG quality (1-100)
        in: query
        name: q
        type: integer
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Preview only supported for images
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Preview image transform
      tags:
      - Variants
...
  /api/v1/assets/{id}/variants:
    get:
      description: Returns all generated variants for the asset's current version,
        plus a <code>rebuilding</code> flag that is true when a variant rebuild job
        is in progress (e.g. after a new version upload). Each variant includes a
        <code>download_url</code> for direct file access.
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ListVariantsResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List asset variants
      tags:
      - Variants
    post:
      consumes:
      - application/json
      description: 'Enqueues a background job to generate a transformed variant of
        the asset''s current version. Supported types and their required params: <ul>
        <li><strong>image_resize</strong> — <code>{"width": N, "height": N, "fit":
        "contain|cover|fill"}</code></li> <li><strong>image_convert</strong> — <code>{"format":
        "jpeg|png|webp|avif"}</code></li> <li><strong>image_crop</strong> — <code>{"x":
        N, "y": N, "width": N, "height": N}</code></li> <li><strong>image_watermark</strong>
        — <code>{"text": "...", "position": "..."}</code></li> <li><strong>image_smart_crop</strong>
        — <code>{"width": N, "height": N}</code> (AI-assisted)</li> <li><strong>image_bg_remove</strong>
        — requires <code>REMOVEBG_API_KEY</code> env var</li> <li><strong>video_transcode</strong>
        — <code>{"format": "mp4", "codec": "h264"}</code></li> <li><strong>video_capture_image</strong>
        — <code>{"time_sec": N}</code></li> </ul> Returns a job ID immediately; poll
        <code>GET /api/v1/assets/:id/variants</code> to check completion. Returns
        409 if a variant rebuild is already in progress.'
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Variant type and transform params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateVariantRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.CreateVariantResponse'
        "400":
          description: Invalid variant type or wrong asset type
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Variant rebuild already in progress
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Asset has no current version
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a variant
      tags:
      - Variants
  /api/v1/assets/{id}/variants/{vid}:
    delete:
      description: Permanently removes the variant record and its stored file. Only
        variants attached to the <em>current</em> version can be deleted manually
        — variants on older versions are removed automatically by the retention purge
        job.
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Variant ID
        in: path
        name: vid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset or variant not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Variant belongs to a previous version
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a variant
      tags:
      - Variants
  /api/v1/assets/{id}/variants/{vid}/file:
    get:
      description: Streams the variant's stored file. Content-Type is derived from
        the file extension. An <code>asset_variant_downloaded</code> audit event is
        recorded (browser image prefetch excluded).
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Variant ID
        in: path
        name: vid
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset, variant, or file not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download variant file
      tags:
      - Variants
  /api/v1/assets/{id}/variants/upload:
    post:
      consumes:
      - multipart/form-data
      description: Accepts a multipart file upload and stores it as a variant of type
        <code>manual</code> on the asset's current version. Unlike transform-generated
        variants, manual variants are <em>not</em> automatically rebuilt when a new
        version is uploaded — they persist across version changes. Use this to attach
        pre-processed or third-party exports (e.g. color-corrected TIFF, print-ready
        PDF).
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: File to attach as a manual variant
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.VariantResponse'
        "400":
          description: file field is required
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Asset has no current version
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a manual variant
      tags:
      - Variants
...
swagger: "2.0"